requirements.txt
aerich==0.9.2
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncclick==8.3.0.3
asyncpg==0.31.0
authlib==1.6.6
bcrypt==4.3.0
blinker==1.9.0
boto3==1.42.22
botocore==1.42.22
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cloudinary==1.44.1
cryptography==46.0.3
dictdiffer==0.9.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-jwt==0.3.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
jmespath==1.0.1
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pypika-tortoise==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
s3transfer==0.16.0
six==1.17.0
starlette==0.50.0
tomli==2.3.0
tortoise-orm==0.25.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn==0.40.0
